Migrate db layer to sqlc-generated queries
Replace the hand-written internal/db/queries.go with sqlc-generated
*.sql.go files. SQL query definitions live in internal/db/queries/*.sql
and are processed by sqlc v1.30.0 to produce type-safe Go wrappers using
the pgx/v5 driver.

What changed:
- Added sqlc.yaml configuration with pgvector and uuid overrides
- Added internal/db/queries/*.sql — one file per domain, SQL extracted
  verbatim from the original queries.go
- Added github.com/pgvector/pgvector-go for the pgvector.Vector type
  used in Embedding/EmbeddingCode fields
- Removed internal/db/queries.go (replaced by generated *.sql.go files)
- internal/db/models.go is now sqlc-generated (was hand-written)
- internal/db/compat.go provides the legacy free-function API
  (CreateMemory, GetMemory, etc.) so all existing callers compile
  without change
- Fixed type mismatches across callers: pgvector.Vector vs []float32,
  int32 vs int, time.Time vs *time.Time, uuid.UUID vs *uuid.UUID,
  ScopeIds casing (sqlc lowercases the field)
- Fixed FindNearDuplicates: added ::float8 cast so sqlc infers threshold
  param as float64 instead of pgvector.Vector

All 17 packages pass go test ./...

+++ cmd/postbrain-hook/main.go
@@ -281,7 +281,7 @@ func runSkillSync(ctx context.Context, agentType, workdir string) error {
 			Description: sk.Description,
 			AgentTypes:  sk.AgentTypes,
 			Body:        sk.Body,
-			Version:     sk.Version,
+			Version:     int32(sk.Version),
 			Parameters:  sk.Parameters,
 		}
